move the coordinator to its own file
