+++ homeassistant/components/electric_kiwi/__init__.py
@@ -14,7 +14,9 @@ from homeassistant.helpers import aiohttp_client, config_entry_oauth2_flow
 from . import api
 from .const import DOMAIN
 
-PLATFORMS: list[Platform] = [Platform.SENSOR]
+PLATFORMS: list[Platform] = [
+    Platform.SENSOR,
+]
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
